Resolve type issue with color prop
The HTMLAttributes prop type defines color as a legacy HTML attribute
that can be set as a prop, this prop definition conflicts with the
styled system definition of color. This issue comes up when trying to
accept the BoxProps on a non-styled-components component and splatting
them to one. This resolves this type issue negating the need for the
various StructuralProps vs StylisticProps.

Note: I tried using the same method with `type BoxProps` however the
issue still remained. We need to extend the types rather than use an
intersection type.