Introduce anchor default translations
Fixes issue with namespace fallbacks.

Originally I was under the impression that when providing multiple
namespaces to the `useTranslation` hook would set up a fallback order
for looking up translations. This is actually incorrect, the first
namespace provided will be the default.

This explodes the array of namespaces to build out a proper fallback
order between custom namespaces, the forms namespace, and a default
anchorForms namespace.