Fix submitting disabled state in button
Previously, I updated the InnerButton to include submitting as a
qualifier for being disabled, this applied the styles to look like the
button was disabled but did not actually disable the button because the
button didn't have disabled applied properly.

This resolves the issue by allowing the button element itself be the
source of truth for whether or not it's disabled and the styled inner
button component uses nesting to properly style itself based off whether
it is within a disabled button.