useTranslatedOptions v2
Instead of returning an array of options, this hook has been updated to
return an object with keys for options and getOptionFromValue. The
options key preserves the original return value of translated options
while getOptionFromValue is a utility function built from the option set
to make it easier to find the option object from a given value.