Allow `TextButton` to act as a link
Similar to the `Button` component, we can now do `<TextButton as="a"
href="...">...</TextButton>` to render a link.

Also, fixes a bug around the font-size of the `TextButton` component.
The `size: "s"` prop applied via `.attrs` wasn't being forwarded to the
underlined `Heading` component so, instead, we now extend from just
`button` and apply the needed font properties directly in the css body.